Moved ImageComment, ImageCommentsItemsMapper, RemoteImageCommentsLoader to their own files which is production ready.
